Module opentelemetry_stackdriver::proto::google::api

Modules

http_rule
Nested message and enum types in HttpRule.

resource_descriptor
Nested message and enum types in ResourceDescriptor.

Structs

CustomHttpPattern
A custom pattern is used for defining custom HTTP verb.

Http
Defines the HTTP configuration for an API service. It contains a list of [HttpRule][google.api.HttpRule], each specifying the mapping of an RPC method to one or more HTTP REST API methods.

HttpRule
gRPC Transcoding

ResourceDescriptor
A simple descriptor of a resource type.

ResourceReference
Defines a proto annotation that describes a string field that refers to an API resource.

Enums

FieldBehavior
An indicator of the behavior of a given field (for example, that a field is required in requests, or given as output but ignored as input). This does not change the behavior in protocol buffers itself; it only denotes the behavior and may affect how API tooling handles the field.